# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digitalna Simulacija TRPG API",
    "version": "1.0.0",
    "description": "Swagger spec za ceo API folder projekta"
  },
  "paths": {
...
    "/api/log-in": {
      "post": {
        "summary": "Prijava korisnika",
        "description": "Autentifikacija korisnika, vraća token u cookie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username","password"],
                "properties": {
                  "username":{"type":"string","example":"john_doe"},
                  "password":{"type":"string","format":"password","example":"123456"}
                }
              }
            }
          }
        },
        "responses": {
          "200":{"description":"Uspesna prijava","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},
          "400":{"description":"Greska u unosu","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},
          "401":{"description":"Neuspesna autentifikacija","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
  },
  "components":{
    "schemas":{
      "User":{"type":"object","properties":{"id":{"type":"integer","example":1},"username":{"type":"string","example":"john_doe"}}},
...
      "ErrorResponse":{"type":"object","properties":{"error":{"type":"string","example":"Nesto nije uredu"}}}
    }
  }
}